refactor

+++ Zadaca1RPR_17324/Zadaca1RPR_17324/PacijentKonstruktori.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Doktori;
 
 namespace Zadaca1RPR_17324
 {
@@ -20,7 +21,7 @@ namespace Zadaca1RPR_17324
             datumPrijema = new DateTime(2017, 1, 1, 0, 0, 0);
             //idPacijenta++;
         }
-        public Pacijent(int maticniBroj1, string ime1, string prezime1, DateTime datumRodjenja1, char spol1, string adresaStanovanja1, string bracnoStanje1, DateTime datumPrijema1)
+        public Pacijent(UInt64 maticniBroj1, string ime1, string prezime1, DateTime datumRodjenja1, char spol1, string adresaStanovanja1, string bracnoStanje1, DateTime datumPrijema1)
         {
             MaticniBroj = maticniBroj1;
             ime = ime1;
